Fix migration

+++ src/Attribute/File.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/attribute_file.
  *
- * (c) 2012-2022 The MetaModels team.
+ * (c) 2012-2023 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -23,7 +23,7 @@
  * @author     David Molineus <david.molineus@netzmacht.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2022 The MetaModels team.
+ * @copyright  2012-2023 The MetaModels team.
  * @license    https://github.com/MetaModels/attribute_file/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -198,6 +198,13 @@ class File extends BaseComplex
      */
     public function destroyAUX()
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
         parent::destroyAUX();
         $metaModel = $this->getMetaModel()->getTableName();
 
@@ -217,6 +224,13 @@ class File extends BaseComplex
      */
     public function initializeAUX()
     {
+        if ($this->isManagedAttribute($this->get('type'))) {
+            $this->triggerDeprecationShouldNotCallManaged(static::class, __METHOD__);
+            return;
+        }
+
+        $this->triggerDeprecationIsUnmanagedAttribute(static::class, __METHOD__);
+
         parent::initializeAUX();
         if ($colName = $this->getColName()) {
             $tableName = $this->getMetaModel()->getTableName();
